requirements.txt
discord.py>=2.3.0
openai>=1.0.0
groq>=0.4.0
python-dotenv>=1.0.0
pytest>=7.0.0
asyncio
aiosqlite>=0.19.0
